Move kafka poll and db storing to its own method, add test
Koden som leser fra kafka og lagrer til db ligger nå utenfor bootstrap, slik at vi kan teste den uten å forholde oss til coroutines
